requirements.txt
ipykernel==6.23.1
numpy==1.24.3
tokenizers==0.13.3
torch
tqdm==4.65.0
transformers
datasets
beautifulsoup4==4.12.2
wandb==0.15.3
hydra-core==1.3.2
tensor-parallel==1.2.4
pyyaml